chat/chat/requirements.txt
fastapi[all]
uvicorn
pyjwt
python-dotenv